DashBoard/pyproject.toml
[project]
name = "proyecto-monitoreo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "aiortc>=1.14.0",
    "av>=16.1.0",
    "cv>=1.0.0",
    "dotenv>=0.9.9",
    "easyocr>=1.7.2",
    "faiss-cpu>=1.13.2",
    "insightface>=0.7.3",
    "numpy>=2.2.6",
    "onnx==1.20.1",
    "onnxruntime-gpu==1.24.2",
    "onvif-zeep>=0.2.12",
    "opencv-python>=4.13.0.92",
    "opencv-python-headless>=4.13.0.92",
    "paho-mqtt>=2.1.0",
    "psycopg>=3.3.3",
    "psycopg-binary>=3.3.3",
    "psycopg-pool>=3.3.0",
    "pyyaml>=6.0.2",
    "ultralytics>=8.4.16",
    "zeep>=4.3.2",
]
